% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refpmsi.R
\name{liste_refpmsi}
\alias{liste_refpmsi}
\title{liste_refpmsi}
\usage{
liste_refpmsi(chemin = path.package("refpmsi"))
}
\arguments{
\item{chemin}{= chemin du package refpmsi::}
}
\value{
liste_refpmsi_tbl = liste des referentiels PMSI en tibble
}
\description{
liste_refpmsi
}
