% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refpmsi.R
\name{refpmsi}
\alias{refpmsi}
\title{refpmsi}
\usage{
refpmsi(referentiel, periodepmsi = "",
  chemin = path.package("refpmsi"))
}
\arguments{
\item{referentiel}{= intitule du referentiel pmsi}

\item{periodepmsi}{= liste des annees pmsi}

\item{chemin}{= chemin du package refpmsi::}
}
\value{
refpmsi_tbl = le referentiel selectionne au format tibble
}
\description{
refpmsi
}
